Atualiza tipo de dados para float

+++ Engine/src/Point3d.cpp
@@ -11,41 +11,41 @@ Point3d::Point3d() :
         x_coord(0), y_coord(0), z_coord(0)
 {}
 
-Point3d::Point3d(double x, double y, double z) :
+Point3d::Point3d(float x, float y, float z) :
         x_coord(x), y_coord (y), z_coord (z)
 {}
 
-double Point3d::getXCoord() const
+float Point3d::getXCoord() const
 {
     return x_coord;
 }
 
-void Point3d::setXCoord(double coord)
+void Point3d::setXCoord(float coord)
 {
     x_coord = coord;
 }
 
-double Point3d::getYCoord() const
+float Point3d::getYCoord() const
 {
     return y_coord;
 }
 
-void Point3d::setYCoord(double coord)
+void Point3d::setYCoord(float coord)
 {
     y_coord = coord;
 }
 
-double Point3d::getZCoord() const
+float Point3d::getZCoord() const
 {
     return z_coord;
 }
 
-void Point3d::setZCoord(double coord)
+void Point3d::setZCoord(float coord)
 {
     z_coord = coord;
 }
 
-void Point3d::setCoord(double x, double y,double z)
+void Point3d::setCoord(float x, float y,float z)
 {
 
     this->setXCoord(x);
@@ -56,7 +56,7 @@ void Point3d::setCoord(double x, double y,double z)
 
 Point3d Point3d::addTo(Point3d point)
 {
-    double tmp_x, tmp_y, tmp_z;
+    float tmp_x, tmp_y, tmp_z;
     tmp_x = x_coord + point.getXCoord();
     tmp_y = y_coord + point.getYCoord();
     tmp_z = z_coord + point.getZCoord();
@@ -68,10 +68,10 @@ Point3d Point3d::addTo(Point3d point)
 
 }
 
-Point3d Point3d::multiplyBy(double k)
+Point3d Point3d::multiplyBy(float k)
 {
 
-    double tmp_x, tmp_y, tmp_z;
+    float tmp_x, tmp_y, tmp_z;
     tmp_x = x_coord * k;
     tmp_y = y_coord * k;
     tmp_z = z_coord * k;
@@ -99,36 +99,36 @@ Point3d & Point3d::operator= (const Point3d & other)
 
 Point3d Point3d::crossProduct(Point3d& u, Point3d &v)
 {
-    double x = (u.getYCoord()*v.getZCoord()) - (u.getZCoord()*v.getZCoord());
-    double y = (u.getXCoord()*v.getZCoord()) - (u.getZCoord()*v.getXCoord());
-    double z = (u.getXCoord()*v.getYCoord()) - (u.getYCoord()*v.getXCoord());
+    float x = (u.getYCoord()*v.getZCoord()) - (u.getZCoord()*v.getZCoord());
+    float y = (u.getXCoord()*v.getZCoord()) - (u.getZCoord()*v.getXCoord());
+    float z = (u.getXCoord()*v.getYCoord()) - (u.getYCoord()*v.getXCoord());
     Point3d p(x, y, z);
 
     return p;
 }
 
 
-void Point3d::incrementXCoordBy(double value)
+void Point3d::incrementXCoordBy(float value)
 {
     x_coord+=value;
 }
-void Point3d::incrementYCoordBy(double value)
+void Point3d::incrementYCoordBy(float value)
 {
     y_coord+=value;
 }
-void Point3d::incrementZCoordBy(double value)
+void Point3d::incrementZCoordBy(float value)
 {
     z_coord+=value;
 }
-void Point3d::decrementXCoordBy(double value)
+void Point3d::decrementXCoordBy(float value)
 {
     x_coord-=value;
 }
-void Point3d::decrementYCoordBy(double value)
+void Point3d::decrementYCoordBy(float value)
 {
     y_coord-=value;
 }
-void Point3d::decrementZCoordBy(double value)
+void Point3d::decrementZCoordBy(float value)
 {
     z_coord-=value;
 }
