+++ Engine/src/Translation.h
@@ -15,7 +15,7 @@
 class Translation: public Transformation {
 public:
 	Translation();
-	Translation(double x, double y, double z);
+	Translation(float x, float y, float z);
 	const Point3d& getPoint() const;
 	void applyTransformation() const;
 	~Translation();
