+++ Engine/src/Triangle.cpp
@@ -13,8 +13,8 @@ Triangle::Triangle() :
 		a(), b(), c() {
 }
 
-Triangle::Triangle(double xA, double yA, double zA, double xB, double yB,
-		double zB, double xC, double yC, double zC) :
+Triangle::Triangle(float xA, float yA, float zA, float xB, float yB,
+		float zB, float xC, float yC, float zC) :
 		a (xA, yA, zA), b(xB, yB, zB), c(xC, yC, zC) {
 }
